Introduce checkers and extractors to better split the applicator
