+++ composer.json
@@ -31,6 +31,7 @@
         "phpstan/phpstan-webmozart-assert": "^1",
         "phpunit/phpunit": "^9.5",
         "polishsymfonycommunity/symfony-mocker-container": "^1.0",
+        "sylius/sylius": "^1.12",
         "sylius-labs/coding-standard": "^4.0",
         "symfony/browser-kit": "^5.4|^6.4",
         "symfony/debug-bundle": "^5.4|^6.4",
