+++ src/Checker/VatRateEligibilityCheckerInterface.php
@@ -1,0 +1,18 @@
+<?php
+
+declare(strict_types=1);
+
+namespace FluxSE\SyliusEUVatPlugin\Checker;
+
+use FluxSE\SyliusEUVatPlugin\Entity\EuropeanChannelAwareInterface;
+use FluxSE\SyliusEUVatPlugin\Entity\VATNumberAwareInterface;
+use Sylius\Component\Core\Model\AddressInterface;
+use Sylius\Component\Core\Model\ChannelInterface;
+
+interface VatRateEligibilityCheckerInterface
+{
+    public function check(
+        AddressInterface&VATNumberAwareInterface       $taxationAddress,
+        ChannelInterface&EuropeanChannelAwareInterface $channel,
+    ): bool;
+}
