+++ src/DependencyInjection/Compiler/OrderItemsBasedStrategyCompilerPass.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace FluxSE\SyliusEUVatPlugin\DependencyInjection\Compiler;
 
+use Symfony\Component\DependencyInjection\Argument\ArgumentInterface;
 use Symfony\Component\DependencyInjection\Compiler\CompilerPassInterface;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\DependencyInjection\Reference;
@@ -13,8 +14,12 @@ final class OrderItemsBasedStrategyCompilerPass implements CompilerPassInterface
     public function process(ContainerBuilder $container): void
     {
         $definition = $container->getDefinition('sylius.taxation.order_items_based_strategy');
-        /** @var array $applicators */
+        /** @var array|ArgumentInterface $applicators */
         $applicators = $definition->getArgument(1);
+        // Sylius >= 1.13
+        if ($applicators instanceof ArgumentInterface) {
+            return;
+        }
         $applicators[] = new Reference('flux_se.sylius_eu_vat_plugin.applicator.order_european_vatnumber_applicator');
         $definition->setArgument(1, $applicators);
     }
