+++ templates/Admin/Channel/europeanChannelForm.html.twig
@@ -1,5 +1,3 @@
-{% set form = settings.form %}
-
 <div class="ui segment">
     <h4 class="ui dividing header">{{ 'flux_se.sylius_eu_vat.ui.european_channel'|trans }}</h4>
     <div class="two fields">
