+++ tests/Application/config/packages/flux_se_sylius_eu_vat.yaml
@@ -1,2 +1,3 @@
 imports:
   - { resource: "@FluxSESyliusEUVatPlugin/config/config.yaml" }
+  - { resource: "@FluxSESyliusEUVatPlugin/config/admin.yaml" }
