+++ tests/Application/templates/bundles/SyliusAdminBundle
@@ -1,0 +1,1 @@
+../../../../templates/SyliusAdminBundle
