--- tests/Application/templates/bundles/SyliusAdminBundle/Common/Form/_address.html.twig
@@ -1,3 +1,0 @@
-{% include '@!SyliusAdmin/Common/Form/_address.html.twig' %}
-
-{% include '@FluxSESyliusEUVatPlugin/Form/_vatNumber.html.twig' %}
