--- tests/Application/templates/bundles/SyliusAdminBundle/Common/_address.html.twig
@@ -1,3 +1,0 @@
-{% include '@!SyliusAdmin/Common/_address.html.twig' %}
-
-{% include '@FluxSESyliusEUVatPlugin/_vatNumber.html.twig' %}
