--- tests/Application/templates/bundles/SyliusAdminBundle/_scripts.html.twig
@@ -1,1 +1,0 @@
-{{ encore_entry_script_tags('admin-entry', null, 'admin') }}
