--- tests/Application/templates/bundles/SyliusAdminBundle/_styles.html.twig
@@ -1,1 +1,0 @@
-{{ encore_entry_link_tags('admin-entry', null, 'admin') }}
