+++ tests/Application/templates/bundles/SyliusShopBundle
@@ -1,0 +1,1 @@
+../../../../templates/SyliusShopBundle
