--- tests/Application/templates/bundles/SyliusShopBundle/Common/Form/_address.html.twig
@@ -1,3 +1,0 @@
-{% include '@!SyliusShop/Common/Form/_address.html.twig' %}
-
-{% include '@FluxSESyliusEUVatPlugin/Form/_vatNumber.html.twig' %}
