--- tests/Application/templates/bundles/SyliusShopBundle/Common/_address.html.twig
@@ -1,3 +1,0 @@
-{% include '@!SyliusShop/Common/_address.html.twig' %}
-
-{% include '@FluxSESyliusEUVatPlugin/_vatNumber.html.twig' %}
