--- tests/Application/templates/bundles/SyliusShopBundle/Taxon/_horizontalMenu.html.twig
@@ -1,2 +1,0 @@
-{% set taxons = resources %}
-{% include '@!SyliusShop/Taxon/_horizontalMenu.html.twig' %}
