--- tests/Application/templates/bundles/SyliusShopBundle/_scripts.html.twig
@@ -1,1 +1,0 @@
-{{ encore_entry_script_tags('shop-entry', null, 'shop') }}
