--- tests/Application/templates/bundles/SyliusShopBundle/_styles.html.twig
@@ -1,1 +1,0 @@
-{{ encore_entry_link_tags('shop-entry', null, 'shop') }}
